third_party/blink/web_tests/http/tests/devtools/elements/shadow/elements-panel-shadow-selection-on-refresh-2-expected.txt

Tests that elements panel preserves selected shadow DOM node on page refresh.


Running: setup

Running: testUserAgentShadowRoot
Page reloaded.
Selected node: '#shadow-root (user-agent)'

Running: testOpenShadowRootChild
Page reloaded.
Selected node: 'input'
